# libs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Diabetes Assistant API
  version: 1.0.0
paths:
  /profiles:
    post:
      summary: Profiles Post
      operationId: profiles_post_profiles_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Profiles Post Profiles Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Profiles Get
      operationId: profiles_get_profiles_get
      parameters:
      - name: telegramId
        in: query
        required: true
        schema:
          type: integer
          title: Telegram Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /reminders:
    get:
      summary: Api Reminders
      operationId: api_reminders_reminders_get
      parameters:
      - name: telegramId
        in: query
        required: true
        schema:
          type: integer
          title: Telegram Id
      - name: id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - type: array
                  items:
                    type: object
                    additionalProperties: true
                - type: object
                  additionalProperties: true
                title: Response Api Reminders Reminders Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Api Reminders Post
      operationId: api_reminders_post_reminders_post
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Api Reminders Post Reminders Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /timezone:
    get:
      summary: Get Timezone
      operationId: get_timezone_timezone_get
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Get Timezone Timezone Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Put Timezone
      operationId: put_timezone_timezone_put
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Timezone'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Put Timezone Timezone Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /profile/self:
    get:
      summary: Profile Self
      operationId: profile_self_profile_self_get
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserContext'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /stats:
    get:
      summary: Get Stats
      operationId: get_stats_stats_get
      parameters:
      - name: telegramId
        in: query
        required: true
        schema:
          type: integer
          title: Telegramid
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayStats'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analytics:
    get:
      summary: Get Analytics
      operationId: get_analytics_analytics_get
      parameters:
      - name: telegramId
        in: query
        required: true
        schema:
          type: integer
          title: Telegramid
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalyticsPoint'
                title: Response Get Analytics Analytics Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user:
    post:
      summary: Create User
      description: Ensure a user exists in the database.
      operationId: create_user_user_post
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebUser'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Create User User Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /history:
    post:
      summary: Post History
      description: Save or update a history record in the database.
      operationId: post_history_history_post
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HistoryRecordSchema-Input'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Post History History Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Get History
      description: Return history records for the authenticated user.
      operationId: get_history_history_get
      parameters:
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryRecordSchema-Output'
                title: Response Get History History Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /history/{record_id}:
    delete:
      summary: Delete History
      description: Delete a history record after verifying ownership.
      operationId: delete_history_history__record_id__delete
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: string
          title: Record Id
      - name: X-Telegram-Init-Data
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Telegram-Init-Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Delete History History  Record Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AnalyticsPoint:
      properties:
        date:
          type: string
          title: Date
        sugar:
          type: number
          title: Sugar
      type: object
      required:
      - date
      - sugar
      title: AnalyticsPoint
    DayStats:
      properties:
        sugar:
          type: number
          title: Sugar
        breadUnits:
          type: number
          title: Breadunits
        insulin:
          type: number
          title: Insulin
      type: object
      required:
      - sugar
      - breadUnits
      - insulin
      title: DayStats
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HistoryRecordSchema-Input:
      properties:
        id:
          type: string
          title: Id
        date:
          type: string
          format: date
          title: Date
        time:
          type: string
          format: time
          title: Time
        sugar:
          anyOf:
          - type: number
          - type: 'null'
          title: Sugar
        carbs:
          anyOf:
          - type: number
          - type: 'null'
          title: Carbs
        breadUnits:
          anyOf:
          - type: number
          - type: 'null'
          title: Breadunits
        insulin:
          anyOf:
          - type: number
          - type: 'null'
          title: Insulin
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        type:
          type: string
          enum:
          - measurement
          - meal
          - insulin
          title: Type
      type: object
      required:
      - id
      - date
      - time
      - type
      title: HistoryRecordSchema
      description: Schema for user history records.
    HistoryRecordSchema-Output:
      properties:
        id:
          type: string
          title: Id
        date:
          type: string
          title: Date
        time:
          type: string
          title: Time
        sugar:
          anyOf:
          - type: number
          - type: 'null'
          title: Sugar
        carbs:
          anyOf:
          - type: number
          - type: 'null'
          title: Carbs
        breadUnits:
          anyOf:
          - type: number
          - type: 'null'
          title: Breadunits
        insulin:
          anyOf:
          - type: number
          - type: 'null'
          title: Insulin
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        type:
          type: string
          enum:
          - measurement
          - meal
          - insulin
          title: Type
      type: object
      required:
      - id
      - date
      - time
      - type
      title: HistoryRecordSchema
      description: Schema for user history records.
    ProfileSchema:
      properties:
        telegramId:
          type: integer
          title: Telegram Id
        icr:
          type: number
          title: Icr
        cf:
          type: number
          title: Cf
        target:
          type: number
          title: Target
        low:
          type: number
          title: Low
        high:
          type: number
          title: High
        orgId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Org Id
      type: object
      required:
      - telegramId
      - icr
      - cf
      - target
      - low
      - high
      title: ProfileSchema
    ReminderSchema:
      properties:
        telegramId:
          type: integer
          title: Telegram Id
        id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id
        type:
          type: string
          title: Type
        time:
          anyOf:
          - type: string
          - type: 'null'
          title: Time
        intervalHours:
          anyOf:
          - type: integer
          - type: 'null'
          title: Interval Hours
        minutesAfter:
          anyOf:
          - type: integer
          - type: 'null'
          title: Minutes After
        isEnabled:
          type: boolean
          title: Is Enabled
          default: true
        orgId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Org Id
      type: object
      required:
      - telegramId
      - type
      title: ReminderSchema
    Timezone:
      properties:
        tz:
          type: string
          title: Tz
      type: object
      required:
      - tz
      title: Timezone
    UserContext:
      properties:
        id:
          type: integer
          title: Id
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        username:
          type: string
          title: Username
        language_code:
          type: string
          title: Language Code
        is_premium:
          type: boolean
          title: Is Premium
      type: object
      required:
      - id
      title: UserContext
      description: Telegram user data supplied via WebApp init data.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WebUser:
      properties:
        telegramId:
          type: integer
          title: Telegram Id
      type: object
      required:
      - telegramId
      title: WebUser