# libs/ts-sdk/models/ProfileSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Diabetes Assistant API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProfileSchema
 */
export interface ProfileSchema {
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    telegramId: number;
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    icr: number;
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    cf: number;
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    target: number;
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    low: number;
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    high: number;
    /**
     * 
     * @type {number}
     * @memberof ProfileSchema
     */
    orgId?: number | null;
}

/**
 * Check if a given object implements the ProfileSchema interface.
 */
export function instanceOfProfileSchema(value: object): value is ProfileSchema {
    if (!('telegramId' in value) || value['telegramId'] === undefined) return false;
    if (!('icr' in value) || value['icr'] === undefined) return false;
    if (!('cf' in value) || value['cf'] === undefined) return false;
    if (!('target' in value) || value['target'] === undefined) return false;
    if (!('low' in value) || value['low'] === undefined) return false;
    if (!('high' in value) || value['high'] === undefined) return false;
    return true;
}

export function ProfileSchemaFromJSON(json: any): ProfileSchema {
    return ProfileSchemaFromJSONTyped(json, false);
}

export function ProfileSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileSchema {
    if (json == null) {
        return json;
    }
    return {
        
        'telegramId': json['telegramId'],
        'icr': json['icr'],
        'cf': json['cf'],
        'target': json['target'],
        'low': json['low'],
        'high': json['high'],
        'orgId': json['orgId'] == null ? undefined : json['orgId'],
    };
}

export function ProfileSchemaToJSON(json: any): ProfileSchema {
    return ProfileSchemaToJSONTyped(json, false);
}

export function ProfileSchemaToJSONTyped(value?: ProfileSchema | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'telegramId': value['telegramId'],
        'icr': value['icr'],
        'cf': value['cf'],
        'target': value['target'],
        'low': value['low'],
        'high': value['high'],
        'orgId': value['orgId'],
    };
}